+++ HRLeaveManagement.Application/Features/LeaveRequests/Handlers/Commands/DeleteLeaveRequestCommandHandler.cs
@@ -1,6 +1,8 @@
 ﻿using AutoMapper;
+using HRLeaveManagement.Application.Exceptions;
 using HRLeaveManagement.Application.Features.LeaveRequests.Requests.Commands;
 using HRLeaveManagement.Application.Persistence.Contracts;
+using HRLeaveManagement.Domain;
 using MediatR;
 
 namespace HRLeaveManagement.Application.Features.LeaveRequests.Handlers.Commands;
@@ -18,6 +20,12 @@ public class DeleteLeaveRequestCommandHandler : IRequestHandler<DeleteLeaveReque
     public async Task<Unit> Handle(DeleteLeaveRequestCommand command, CancellationToken cancellationToken)
     {
         var leaveRequest = await _leaveRequestRepository.Get(command.Id);
+
+        if (leaveRequest == null)
+        {
+            throw new NotFoundException(nameof(LeaveRequest), command.Id);
+        }
+
         await _leaveRequestRepository.Delete(leaveRequest);
         return Unit.Value;
     }
